Refactor k-calendar
