+++ panel/src/ui/components/Calendar.vue
@@ -1,46 +1,70 @@
 <template>
-  <div class="k-calendar-input">
+  <div class="k-calendar-input bg-black rounded-sm">
     <nav>
-      <k-button icon="angle-left" @click="prev" />
+      <k-button
+        icon="angle-left"
+        @click="prev"
+      />
       <span class="k-calendar-selects">
         <k-select-input
+          v-model.number="month"
           :options="months"
           :disabled="disabled"
           :required="true"
-          v-model.number="month"
         />
         <k-select-input
+          v-model.number="year"
           :options="years"
           :disabled="disabled"
           :required="true"
-          v-model.number="year"
         />
       </span>
-      <k-button icon="angle-right" @click="next" />
+      <k-button
+        icon="angle-right"
+        @click="next"
+      />
     </nav>
     <table class="k-calendar-table">
       <thead>
         <tr>
-          <th v-for="day in weekdays" :key="'weekday_' + day">{{ day }}</th>
+          <th
+            v-for="dayHeader in weekdays"
+            :key="'weekday_' + dayHeader"
+          >
+            {{ dayHeader }}
+          </th>
         </tr>
       </thead>
       <tbody>
-        <tr v-for="week in numberOfWeeks" :key="'week_' + week">
+        <tr
+          v-for="week in numberOfWeeks"
+          :key="'week_' + week"
+        >
           <td
-            v-for="(day, dayIndex) in days(week)"
+            v-for="(dayButton, dayIndex) in days(week)"
             :key="'day_' + dayIndex"
-            :aria-current="isToday(day) ? 'date' : false"
-            :aria-selected="isCurrent(day) ? 'date' : false"
+            :aria-current="isToday(dayButton) ? 'date' : false"
+            :aria-selected="isCurrent(dayButton) ? 'date' : false"
             class="k-calendar-day"
           >
-            <k-button v-if="day" @click="select(day)">{{ day }}</k-button>
+            <k-button
+              v-if="dayButton"
+              :text="dayButton"
+              @click="select(dayButton)"
+            />
           </td>
         </tr>
       </tbody>
       <tfoot>
         <tr>
-          <td class="k-calendar-today" colspan="7">
-            <k-button @click="selectToday">{{ $t("today") }}</k-button>
+          <td
+            class="k-calendar-today"
+            colspan="7"
+          >
+            <k-button
+              :text="$t('today')"
+              @click="selectToday"
+            />
           </td>
         </tr>
       </tfoot>
@@ -52,11 +76,13 @@
 
 export default {
   props: {
+    /**
+     * ISO date string. I.e. 2012-12-12
+     */
     value: String,
     disabled: Boolean
   },
   data() {
-
     const current = this.value ? this.$library.dayjs(this.value) : this.$library.dayjs();
 
     return {
@@ -223,9 +249,7 @@ $cell-padding: 0.25rem 0.5rem;
 
 .k-calendar-input {
   padding: 0.5rem;
-  background: $color-black;
   color: $color-light;
-  border-radius: $rounded-xs;
 }
 .k-calendar-table {
   table-layout: fixed;
